requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.2
tiktoken==0.7.0
firebase-admin==6.5.0
google-cloud-firestore==2.13.1
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
python-dateutil==2.9.0.post0
twilio==9.3.3
apscheduler==3.10.4
chromadb==0.5.5